Large readability reorg.
