--- saylua/modules/explore/static-source/es6/Dungeon/Actions/GameState.js
@@ -1,6 +1,0 @@
-export const movePlayer = (direction) => {
-  return {
-    type: 'MOVE_PLAYER',
-    direction
-  };
-};
