+++ saylua/modules/explore/static-source/es6/Dungeon/Utils/engine.js
@@ -1,4 +1,6 @@
-import { OBSTRUCTIONS } from "./game_helpers";
+// engine
+// --------------------------------------
+// Primarily relates to app level functions outside of the logic of the game itself.
 
 
 export function log(message) {
@@ -21,42 +23,4 @@ export function log(message) {
   console.log(`Logging: ${output}`);
   window.logQueue.push(output);
   return output;
-}
-
-export function generateNodeGraph(tileSet, tileLayer) {
-  let nodeGraph = [];
-
-  // Our A* implementation uses [x][y] grids, so we must convert from our [y][x] grids.
-  // Weight based on whether or not they are obstructions.
-  for (let tile of tileLayer) {
-    nodeGraph[tile.location.x] = nodeGraph[tile.location.x] || [];
-
-    let parentTileType = tileSet[tile.tile].type;
-    let isObstruction = (OBSTRUCTIONS.indexOf(parentTileType) !== -1);
-
-    if (isObstruction) {
-      nodeGraph[tile.location.x].push(0);
-    } else {
-      nodeGraph[tile.location.x].push(1);
-    }
-  }
-
-  return nodeGraph;
-}
-
-export function normalizeTileLayer(tileLayer) {
-  let mapHeight = tileLayer.length;
-  let mapWidth = tileLayer[0].length;
-  let normalizedTileLayer = [];
-
-  for (let y = 0; y < mapHeight; y++) {
-    for (let x = 0; x < mapWidth; x++) {
-      let tile = tileLayer[y][x];
-
-      tile.location = { x, y };
-      normalizedTileLayer.push(tile);
-    }
-  }
-
-  return [mapHeight, mapWidth, normalizedTileLayer];
 }
